requirements.txt
google-api-python-client==2.143.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
trafilatura==1.12.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
